Add RuntimeOptions
