+++ crates/swc/src/lib.rs
@@ -221,8 +221,6 @@ pub struct Compiler {
     /// CodeMap
     pub cm: Arc<SourceMap>,
     comments: SwcComments,
-    #[cfg(feature = "plugin")]
-    plugin_runtime: Option<Arc<dyn swc_plugin_runner::runtime::Runtime>>,
 }
 
 /// These are **low-level** apis.
@@ -464,20 +462,7 @@ impl Compiler {
         Compiler {
             cm,
             comments: Default::default(),
-            #[cfg(all(feature = "plugin", feature = "plugin_backend_wasmer"))]
-            plugin_runtime: Some(Arc::new(swc_plugin_backend_wasmer::WasmerRuntime)),
-            #[cfg(all(feature = "plugin", not(feature = "plugin_backend_wasmer")))]
-            plugin_runtime: None,
         }
-    }
-
-    #[cfg(feature = "plugin")]
-    pub fn plugin_runtime(
-        mut self,
-        plugin_runtime: Arc<dyn swc_plugin_runner::runtime::Runtime>,
-    ) -> Self {
-        self.plugin_runtime = Some(plugin_runtime);
-        self
     }
 
     #[tracing::instrument(skip_all)]
@@ -643,8 +628,6 @@ impl Compiler {
                 Some(config),
                 comments,
                 before_pass,
-                #[cfg(feature = "plugin")]
-                self.plugin_runtime.clone(),
             )?;
             Ok(Some(built))
         })
